Use source file mapping for all compilation providers
All compilation providers should be setting the source file attribute
for use in debugging, there is no need to limit this to Java. This should
allow for deletions of Kotlin classes to be handled correctly.
